backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile backend/requirements.in
#
click==8.0.4
    # via flask
flask==2.0.3
    # via
    #   -r backend/requirements.in
    #   flask-sqlalchemy
flask-sqlalchemy==2.5.1
    # via -r backend/requirements.in
greenlet==1.1.2
    # via sqlalchemy
itsdangerous==2.1.0
    # via flask
jinja2==3.0.3
    # via flask
markupsafe==2.1.0
    # via jinja2
psycopg2==2.9.3
    # via -r backend/requirements.in
pyyaml==6.0
    # via -r backend/requirements.in
sqlalchemy==1.4.32
    # via flask-sqlalchemy
werkzeug==2.0.3
    # via flask
